removed next boilerplate index
